cleanedup code
